Cleaned a code + directories. Added small styles in banner.
