--- src/components/Header/AboutMe.scss
@@ -1,9 +1,0 @@
-@import "../../variables";
-
-.section-about-me {
-  height: $height-header;
-  //background-color: #123;
-  position: relative;
-}
-
-
