--- src/components/Header/AboutMe.tsx
@@ -1,13 +1,0 @@
-import React from 'react';
-import { Banner } from './Banner/Banner';
-
-import './AboutMe.scss';
-
-export const AboutMe = () => {
-
-    return (
-        <section className="section-about-me">
-            <Banner/>
-        </section>
-    )
-}
