+++ src/components/Sections/Tech/Tech.scss
