+++ src/views/MainPageView.tsx
@@ -1,5 +1,4 @@
 import React, {useEffect} from 'react';
-import {AboutMe} from "../components/Header/AboutMe";
 import {Sections} from "../components/Sections/Sections";
 import {DataResponseType} from 'types';
 import {useDispatch, useSelector} from "react-redux";
@@ -10,29 +9,26 @@ import {LoadingView} from "./LoadingView/LoadingView";
 export const MainPageView = () => {
   const {projects, contacts} = useSelector((store: RootState) => store.cvData)
   const dispatch = useDispatch();
-    useEffect(() => {
-      (async () => {
-        const res = await fetch('http://localhost:3001/api');
-        const data: DataResponseType = await res.json();
-        console.log(data)
-        if (data.success) {
-          dispatch(setProjects(data.projects.projects))
-          dispatch(setContacts(data.contacts.contacts))
+  useEffect(() => {
+    (async () => {
+      const res = await fetch('http://localhost:3001/api');
+      const data: DataResponseType = await res.json();
+      console.log(data)
+      if (data.success) {
+        dispatch(setProjects(data.projects.projects))
+        dispatch(setContacts(data.contacts.contacts))
 
-        } else {
-          window.location.href = 'http://localhost:3000/error'
-        }
-      })()
-    }, [])
+      } else {
+        window.location.href = 'http://localhost:3000/error'
+      }
+    })()
+  }, [])
 
-  if (projects.length === 0 && contacts.length === 0){
+  if (projects.length === 0 && contacts.length === 0) {
     return <LoadingView/>
   }
 
-    return (
-        <main>
-            <AboutMe/>
-            <Sections/>
-        </main>
-    )
+  return (
+    <Sections/>
+  )
 }
